Modal Edit
Today I figured out more of Blenders functions and controls, used it's 'Sculpting' Tools to give the basic modal a face, and got a lot of its skeletal mesh created. Would have been done on schedule but discovered  a 'restraint' function in creating the bones that sets how they should bend when it comes to animating them. This should be done tomorrow. Then work can start on giving this personally made modal a walk/run animation and turn it into the player character for future development.